fixing single quote frontend page

+++ calc/models.py
@@ -11,8 +11,8 @@ LAYOUT = [('1', '1 page'), ('2', '2 pages'), ('3', '3 pages'), ('4', '4 page'), 
 class Input(models.Model):
     partner_name = models.ForeignKey('auth.User', on_delete=models.CASCADE)
     customer_name = models.CharField(max_length=40, verbose_name='Client')
-    offline_data = models.FloatField(verbose_name='Offline data (GB)', default=0)
-    online_data = models.FloatField(verbose_name='Online data (GB)', default=0)
+    offline_data = models.IntegerField(verbose_name='Offline data (GB)', default=0)
+    online_data = models.IntegerField(verbose_name='Online data (GB)', default=0)
     type = models.CharField(default='digital', max_length=40, choices=TYPE, verbose_name='Type of preservation', blank=True )
     pages = models.IntegerField(default=0, blank=True, verbose_name='How many pages')
     layout = models.CharField(default='1', blank=True, max_length=40, choices=LAYOUT, verbose_name='Pages per frame')
